Update to make store by category
