+++ src/main/java/com/example/demo/src/store/StoreProvider.java
@@ -1,7 +1,7 @@
 package com.example.demo.src.store;
 
 import com.example.demo.config.BaseException;
-import com.example.demo.src.store.model.GetCategoryRes;
+import com.example.demo.src.store.model.GetStoreRes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,47 +11,46 @@ import java.util.List;
 
 import static com.example.demo.config.BaseResponseStatus.*;
 
-
 @Service
 
-public class CategoryProvider {
+public class StoreProvider {
 
     final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    private final CategoryDao categoryDao;
+    private final StoreDao storeDao;
 
     @Autowired
-    public CategoryProvider(CategoryDao categoryDao) {
-        this.categoryDao = categoryDao;
+    public StoreProvider(StoreDao storeDao) {
+        this.storeDao = storeDao;
     }
 
     // ************************************************************************************
 
     // 최신순 정렬
-    public List<GetCategoryRes> getCategoryByDate(int userIdx, int categoryIdx, String deliveryService, int pageNo) throws BaseException {
+    public List<GetStoreRes> getCategoryByDate(int userIdx, int categoryIdx, String deliveryService, int pageNo) throws BaseException {
         try {
-            List<GetCategoryRes> getCategoryRes;
+            List<GetStoreRes> getCategoryRes;
 
             // category => 전체보기
-            if (categoryDao.checkCategoryIdx(categoryIdx) == 1) {
+            if (storeDao.checkCategoryIdx(categoryIdx) == 1) {
                 // 전체보기 - 최신순 정렬 + 배달 가능 설정 안 함
                 if (deliveryService.equals("N")) {
-                    getCategoryRes = categoryDao.getAllByDate(userIdx, pageNo);
+                    getCategoryRes = storeDao.getAllByDate(userIdx, pageNo);
                 }
                 // 전체보기 - 최신순 정렬 + 배달 가능
                 else {
-                    getCategoryRes = categoryDao.getAllByDateAndDelivery(userIdx, deliveryService, pageNo);
+                    getCategoryRes = storeDao.getAllByDateAndDelivery(userIdx, deliveryService, pageNo);
                 }
             }
             // category => 한식 ..
             else {
                 // 카테고리 - 최신순 정렬 + 배달 가능 설정 안 함
                 if (deliveryService.equals("N")) {
-                    getCategoryRes = categoryDao.getCategoryByDate(userIdx, categoryIdx, pageNo);
+                    getCategoryRes = storeDao.getCategoryByDate(userIdx, categoryIdx, pageNo);
                 }
                 // 카테고리 - 최신순 정렬 + 배달 가능
                 else {
-                    getCategoryRes = categoryDao.getCategoryByDateAndDelivery(userIdx, categoryIdx, deliveryService, pageNo);
+                    getCategoryRes = storeDao.getCategoryByDateAndDelivery(userIdx, categoryIdx, deliveryService, pageNo);
                 }
             }
             return getCategoryRes;
@@ -60,31 +59,32 @@ public class CategoryProvider {
         }
     }
 
+
     // 별점순 정렬
-    public List<GetCategoryRes> getCategoryByStarRate(int userIdx, int categoryIdx, String deliveryService, int pageNo) throws BaseException {
+    public List<GetStoreRes> getCategoryByStarRate(int userIdx, int categoryIdx, String deliveryService, int pageNo) throws BaseException {
         try {
-            List<GetCategoryRes> getCategoryRes;
+            List<GetStoreRes> getCategoryRes;
 
             // category => 전체보기
-            if (categoryDao.checkCategoryIdx(categoryIdx) == 1) {
+            if (storeDao.checkCategoryIdx(categoryIdx) == 1) {
                 // 전체보기 - 별점순 정렬 + 배달 가능 설정 안 함
                 if (deliveryService.equals("N")) {
-                    getCategoryRes = categoryDao.getAllByStarRate(userIdx, pageNo);
+                    getCategoryRes = storeDao.getAllByStarRate(userIdx, pageNo);
                 }
                 // 전체보기 - 별점순 정렬 + 배달 가능
                 else {
-                    getCategoryRes = categoryDao.getAllByStarRateAndDelivery(userIdx, deliveryService, pageNo);
+                    getCategoryRes = storeDao.getAllByStarRateAndDelivery(userIdx, deliveryService, pageNo);
                 }
             }
             // category => 한식 ..
             else {
                 // 카테고리 - 최신순 정렬 + 배달 가능 설정 안 함
                 if (deliveryService.equals("N")) {
-                    getCategoryRes = categoryDao.getCategoryByStarRate(userIdx, categoryIdx, pageNo);
+                    getCategoryRes = storeDao.getCategoryByStarRate(userIdx, categoryIdx, pageNo);
                 }
                 // 카테고리 - 최신순 정렬 + 배달 가능
                 else {
-                    getCategoryRes = categoryDao.getCategoryByStarRateAndDelivery(userIdx, categoryIdx, deliveryService, pageNo);
+                    getCategoryRes = storeDao.getCategoryByStarRateAndDelivery(userIdx, categoryIdx, deliveryService, pageNo);
                 }
             }
             return getCategoryRes;
